refactor(bin-designer): use shared usePreviewColor hook in CompartmentEditor
Address review feedback on PR #1528:

- Drop the duplicated color helpers in compartmentEditorColors.ts; use
  the existing usePreviewColor() hook + getPreviewBorderColor() and add
  getCompartmentFill() to the shared module so behavior stays
  centralized (Copilot).
- Switch CompartmentEditorParts.tsx to a named ReactNode import
  instead of relying on the React global namespace (Greptile P2).
- Replace the misleading "any cell in this compartment is hovered"
  comment — the flag actually keys off this cell's isHovered prop
  (Copilot).
